added Employee and Project Exceptions

+++ src/main/java/com/example/finalProject/services/DepartmentService.java
@@ -8,8 +8,6 @@ import com.example.finalProject.repositories.DepartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.NoSuchElementException;
 
 @Service
@@ -18,16 +16,24 @@ public class DepartmentService {
     @Autowired
     private DepartmentRepository departmentRepository;
 
-    //???
     @Autowired
     private EmployeeService employeeService;
 
+    /**
+     * Finds all Departments
+     * @return List of found Departments
+     */
     public Iterable<Department> findAll(){
         return departmentRepository.findAll();
     }
 
+    /**
+     * Search for Department with this id. If not found returns an error
+     * @param id - id of the Department
+     * @return Found Department object
+     */
     public Department findById(Long id){
-        if(id.equals(null) || id <= 0L){
+        if(id == null || id <= 0L){
             throw new WrongArgumentsException();
         }
 
@@ -38,6 +44,11 @@ public class DepartmentService {
         }
     }
 
+    /**
+     * Saves new Department in the DB. First checks if all properties are valid
+     * @param department - new Department object to saved
+     * @return - new Department object saved in a DB with id from the DB
+     */
     public Department save(Department department){
         if(department.getName() == null || department.getName().trim().isEmpty()){
             throw new WrongArgumentsException();
@@ -45,6 +56,12 @@ public class DepartmentService {
         return departmentRepository.save(department);
     }
 
+    /**
+     * Updates already existing Department
+     * @param department - new Department object
+     * @param id - id of the Department that to be updated
+     * @return - updated Department object
+     */
     public Department update (Department department, Long id){
         Department foundDepartment = findById(id);
         if(department.getName() == null || department.getName().trim().isEmpty()){
@@ -55,14 +72,32 @@ public class DepartmentService {
         return foundDepartment;
     }
 
+    /**
+     * Deletes Department object by id
+     * @param id - id of the Department to be deleted
+     */
     public void deleteById(Long id){
+        if(id <= 0L){
+            throw new WrongArgumentsException();
+        }
         departmentRepository.deleteById(id);
     }
 
+    /**
+     * Finds all Employees in a Department
+     * @param depId - id of the Department
+     * @return - Employee object with all Employees in this Department
+     */
     public Iterable<Employee> allEmployeesInDepartment(Long depId){
+        Department dep = findById(depId);
         return employeeService.allEmployeesInDepartment(depId);
     }
 
+    /**
+     * Sum of salaries in a given department
+     * @param id - id of the Department
+     * @return - Sum of the salaries. If there are no Employees return 0.0
+     */
     public double salarySum(Long id){
         Department department = findById(id);
